\name{read.fortunes}
\alias{read.fortunes}
\title{Read the fortunes file}
\usage{
  read.fortunes(file = NULL)
}
\arguments{
  \item{file}{A character string giving a fortune database
  in csv format (in UTF-8 encoding).  By default all csv
  files in the data directory of the fortune package are
  used.}
}
\description{
  Read the fortunes file
}

